chore: added auth to routes and update catch error
