+++ src/app/api/auth/register/validators.ts
